refactored the over-generic _utils.py code (a mix of helper functions, policy checklist parsing/scoring functions, figure generators, and reporting-related functions) to extract policy-specific functions to the policy_report.py script.  This is a start at reducing the size of this file.  Ideally the report template production functions would be in their own module, and the plotting functions located in _plots.py --- tangentially relates to some discussions in #507 (a start towards more logical structure of code).  But I'm focusing on policy today! --- and updated pre-commit config yaml python language to Python 3.12 (no longer working; configuration issues, will run later)
