+++ .pre-commit-config.yaml
@@ -28,7 +28,7 @@ repos:
     rev: 24.4.2
     hooks:
     -    id: black
-         language_version: python3.9
+         language_version: python3.12
 -   repo: https://github.com/pycqa/flake8
     rev: 7.0.0
     hooks:
